+++ src/lib.rs
@@ -1,10 +1,11 @@
+mod audio_player;
 mod play_sound_event;
 mod plugin;
 mod sync_wrapped;
+mod utils;
 
+pub use crate::audio_player::AudioListener;
+pub use crate::audio_player::AudioSource;
+pub use crate::audio_player::AudioSourcePlayer;
 pub use crate::play_sound_event::PlaySoundEvent;
-pub use crate::plugin::AudioListener;
-pub use crate::plugin::AudioSource;
-pub use crate::plugin::AudioSourcePlayer;
 pub use crate::plugin::FmodPlugin;
-pub use libfmod::Studio; // Todo: We should probably avoid exposing this
